refactored repro scripts
