--- src/main/resources/reproduce/msmarco-v2.1-doc-segmented.yaml
@@ -1,39 +1,0 @@
-conditions:
-  - name: bm25
-    display: "BM25 v2.1 Segmented Corpus (k1=0.9, b=0.4)"
-    display_html: "BM25 v2.1 Segmented Corpus (<i>k<sub><small>1</small></sub></i>=0.9, <i>b</i>=0.4)"
-    display_row: ""
-    command: bin/run.sh io.anserini.search.SearchCollection -threads $threads -index msmarco-v2.1-doc-segmented -topics $topics -output $output -hits 1000 -bm25 -selectMaxPassage -selectMaxPassage.delimiter "#" -selectMaxPassage.hits 1000
-    topics:
-      - topic_key: msmarco-v2-doc-dev
-        eval_key: tools/topics-and-qrels/qrels.msmarco-v2.1-doc.dev.txt
-        scores:
-          - MRR@10: 0.1973
-      - topic_key: msmarco-v2-doc-dev2
-        eval_key: tools/topics-and-qrels/qrels.msmarco-v2.1-doc.dev2.txt
-        scores:
-          - MRR@10: 0.2000
-      - topic_key: trec2021-dl
-        eval_key: tools/topics-and-qrels/qrels.dl21-doc-msmarco-v2.1.txt
-        scores:
-          - MAP: 0.2609
-            MRR@10: 0.9026
-            nDCG@10: 0.5778
-            R@100: 0.3811
-            R@1K: 0.7115
-      - topic_key: trec2022-dl
-        eval_key: tools/topics-and-qrels/qrels.dl22-doc-msmarco-v2.1.txt
-        scores:
-          - MAP: 0.1079
-            MRR@10: 0.7213
-            nDCG@10: 0.3576
-            R@100: 0.2330
-            R@1K: 0.4790
-      - topic_key: trec2023-dl
-        eval_key: tools/topics-and-qrels/qrels.dl23-doc-msmarco-v2.1.txt
-        scores:
-          - MAP: 0.1391
-            MRR@10: 0.6519
-            nDCG@10: 0.3356
-            R@100: 0.3049
-            R@1K: 0.5852
